break out exp projection fns
